Intermediate commit

+++ org.eclipse.xtext.ui/src/org/eclipse/xtext/ui/editor/contentassist/ContentAssistContext.java
@@ -197,7 +197,7 @@ public class ContentAssistContext {
             this.replaceContextLength = replaceContextLength;
             return replaceContextLength;
         }
-        return replaceContextLength.intValue();
+        return replaceContextLength;
     }
 
     /**
